Extract Muni-specific "route heuristics"
