+++ frontend/src/helpers/routeCalculations.js
@@ -10,110 +10,8 @@ import red from '@material-ui/core/colors/red';
 import yellow from '@material-ui/core/colors/yellow';
 import lightGreen from '@material-ui/core/colors/lightGreen';
 import green from '@material-ui/core/colors/green';
-import {
-  getTripTimesForDirection,
-  getAverageOfMedianWait,
-} from './precomputed';
-
-/**
- * Returns a data object with centralized declarations of "per route" heuristic rules
- * to apply when doing systemwide computations.
- *
- * For example, for routes with directions that should be ignored:
- *
- * {
- *   <routeId>: {
- *     directionsToIgnore: [<directionId>]
- *   }
- * }
- *
- * Other cases:
- * - Routes to filter out completely:
- *   - S due to lack of regular route and schedule
- *   - Owls due to the date boundary problem.
- * - Routes that have non-code issues with arrivals their first or last stop and so the second or penultimate stop
- *     should be used instead for end-to-end calculations.  Cable car lines are like this.  Also the M has a last
- *     stop that it normally does not go to.
- * - Possibly special handling for routes with back end issues (currently 5, 9, 9R) as a temporary workaround.
- *   - The 9 has multiple terminals so use the last common stop.
- *   - The 5 was reconfigured and Nextbus stop configs are out of sync with historic data.  Use last good stop.
- */
-export const routeHeuristics = {
-  J: {
-    directionsToIgnore: ['J____I_D10'], // this is to 23rd and 3rd
-  },
-  L: {
-    directionsToIgnore: ['L____I_U53'],
-  },
-  M: {
-    M____O_D00: {
-      ignoreFirstStop: true, // Embarcadero & Folsom is not a real stop
-    },
-  },
-  N: {
-    N____O_F10: {
-      ignoreFirstStop: true, // 4th and King to 2nd and King trip times are skewed by a few hyperlong trips
-    },
-  },
-  S: {
-    ignoreRoute: true,
-  },
-  '5': {
-    '5____I_F00': {
-      ignoreFirstStop: '4218', // no data for 3927, and first few stop ids are now different.  Problem is even worse on outbound side, no good fix there.
-    },
-  },
-  '9': {
-    '9____I_N00': {
-      ignoreFirstStop: '7297', // use Bayshore as actual first stop (daytime)
-    },
-    '9____O_N00': {
-      ignoreLastStop: '7297', // use Bayshore as actual terminal (daytime)
-    },
-  },
-  '24': {
-    directionsToIgnore: ['24___I_D10'],
-  },
-  '90': {
-    ignoreRoute: true,
-  },
-  '91': {
-    ignoreRoute: true,
-  },
-  K_OWL: {
-    ignoreRoute: true,
-  },
-  L_OWL: {
-    ignoreRoute: true,
-  },
-  M_OWL: {
-    ignoreRoute: true,
-  },
-  N_OWL: {
-    ignoreRoute: true,
-  },
-  T_OWL: {
-    ignoreRoute: true,
-  },
-  PM: {
-    PM___O_F00: {
-      ignoreLastStop: true, // long time to Taylor and Bay (probably in holding area)
-    },
-    PM___I_F00: {
-      ignoreFirstStop: true, // 30 minutes from Hyde & Beach to Hyde & North Point
-    },
-  },
-  PH: {
-    PH___I_F00: {
-      ignoreFirstStop: true, // 30 minutes from Hyde & Beach to Hyde & North Point
-    },
-  },
-  C: {
-    C____I_F00: {
-      ignoreLastStop: true, // long time to California & Drumm (probably in holding area)
-    },
-  },
-};
+import { getAverageOfMedianWait, getTripTimesForDirection } from './precomputed';
+import { routeHeuristics } from '../locationConstants';
 
 /**
  * Given an array of routes, return only the routes we want to show.
